# padroes-rest/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3003",
  "basePath": "/",
  "tags": [
    {
      "name": "Todo",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/ping": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/todo/": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Get all todo.",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "todoList": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Todo"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newTodo",
            "in": "body",
            "description": "Informações do todo.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddTodo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Todo criado.",
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          },
          "400": {
            "description": "Invalid Parameters",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "author or description is missing"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/todo/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Get a specific todo.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Todo Id"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Todo"
            },
            "description": "Todo found."
          },
          "400": {
            "description": "Malformed id.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "malformed id"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "todo not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Todo Id"
          },
          {
            "name": "newStatus",
            "in": "body",
            "description": "New todo status.",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "doing"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Todo"
            },
            "description": "Status updated."
          },
          "400": {
            "description": "Malformed id.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "malformed id"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Invalid status.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "the status dong is invalid, please put a valid status e.g \"todo\", \"doing\", \"done\""
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Todo Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful todo deletion."
          },
          "400": {
            "description": "Malformed id.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "malformed id"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "todo"
        },
        "author": {
          "type": "string",
          "example": "testauthor"
        },
        "description": {
          "type": "string",
          "example": "testdescription"
        },
        "id": {
          "type": "string",
          "example": "6112ed7bd38071423dcd6f71"
        }
      }
    },
    "AddTodo": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string",
          "example": "testauthor"
        },
        "description": {
          "type": "string",
          "example": "testdescription"
        }
      },
      "required": [
        "author",
        "description"
      ]
    }
  }
}